requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
mysql-connector-python==9.0.0
pydantic==2.8.2
python-dotenv==1.0.1
tabulate==0.9.0
colorama==0.4.6